add gemsFonction.py
