+++ multimedia/youtube.py
@@ -2,7 +2,7 @@ import os
 import html
 import googleapiclient.discovery
 
-token_youtube = open("vocal/token_youtube.txt","r").read().replace("\n","")
+token_youtube = open("multimedia/token_youtube.txt","r").read().replace("\n","")
 
 def search_youtube(user_input, number):
     # Disable OAuthlib's HTTPS verification when running locally.
